Refactor implementation into own file
